+++ pkg/pve/core_http.go
@@ -2,6 +2,7 @@ package pve
 
 import (
 	"bytes"
+	"context"
 	"crypto/tls"
 	"encoding/json"
 	"fmt"
@@ -175,6 +176,97 @@ func (c *httpClient) sendReq2(method, path string, payload any, result any) erro
 	if err != nil {
 		return err
 	}
+
+	if err := c.Creds.Set(req); err != nil {
+		return err
+	}
+
+	if c.ServiceToken != nil {
+		c.ServiceToken.Set(req)
+	}
+
+	res, err := c.Client.Do(req)
+	if err != nil {
+		return err
+	}
+
+	b, err := io.ReadAll(res.Body)
+	if err != nil {
+		return err
+	}
+
+	if res.StatusCode != http.StatusOK {
+		if !c.APIWrapper {
+			return errors.NewHTTPErrorFromResponse(res)
+		}
+
+		httpErr := new(errors.HTTPError)
+		if err := json.Unmarshal(b, httpErr); err != nil {
+			return err
+		}
+
+		return httpErr
+
+	}
+
+	switch t := reflect.TypeOf(result); t {
+	case reflect.TypeFor[*string]():
+		pveRes := &pveResponse[string]{}
+		err = json.Unmarshal(b, pveRes)
+		if err != nil {
+			return err
+		}
+		*result.(*string) = pveRes.Data
+		return nil
+
+	case reflect.TypeFor[*int]():
+		pveRes := &pveResponse[int]{}
+		err = json.Unmarshal(b, pveRes)
+		if err != nil {
+			return err
+		}
+		*result.(*int) = pveRes.Data
+		return nil
+	default:
+		pveRes := &pveResponse[any]{}
+		err = json.Unmarshal(b, pveRes)
+		if err != nil {
+			return err
+		}
+		b, _ := json.Marshal(pveRes.Data)
+		json.Unmarshal(b, result)
+		return nil
+	}
+}
+
+// sendReq3 sends an http request to the configured proxmox api.
+//
+// It stores the response value into the result parameter only
+// if no error has been returned. If an error is returned, the
+// passed result parameter will be intact.
+func (c *httpClient) sendReq3(
+	method string,
+	path string,
+	payload any,
+	morePayload map[string]string,
+	result any,
+) error {
+	url := c.buildRequestUrl(path)
+
+	req, err := httpin.NewRequest(method, url, payload, httpin.Option.WithNestedDirectivesEnabled(true))
+	if err != nil {
+		return err
+	}
+
+	reqClone := req.Clone(context.Background())
+	reqClone.ParseForm()
+
+	for k, v := range morePayload {
+		reqClone.Form.Add(k, v)
+	}
+
+	newBody := io.NopCloser(strings.NewReader(reqClone.Form.Encode()))
+	req.Body = newBody
 
 	if err := c.Creds.Set(req); err != nil {
 		return err